admin/requirements.txt
django==4.2.27
django-unfold==0.67.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1
httpx==0.28.1
whitenoise==6.7.0
